refactor: move HTTP_REQ_TIMEOUT to eodag.utils
